backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "django-filter>=23.5",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "django-ratelimit>=4.1.0",
    "sentry-sdk>=2.0.0",
    "django-prometheus>=2.3.1",
    "pandas>=2.3.3",
    "ipykernel>=7.1.0",
]
